+++ RELEASES.md
@@ -3,6 +3,7 @@ Version 0.9.2 (2023-03-XX)
 * New driving model, trained on a new dataset
 * Draw MPC path instead of model predicted path, this is a more accurate representation of what the car will do.
 * Buick LaCrosse 2017-19 support thanks to koch-cf!
+* Kia Carnival (China only) 2023 support thanks to sunnyhaibin!
 * Škoda Fabia 2022-23 support thanks to jyoung8607!
 
 Version 0.9.1 (2023-02-28)